添加边播放，下面预览条跟随滚动的功能
还存在BUG， 而且解码90帧时间太长，2400ms左右，需要查看原因

+++ myplayer/src/main/cpp/KzgFFmpeg.h
@@ -58,7 +58,7 @@ private:
     pthread_mutex_t init_mutex;
     pthread_mutex_t seek_mutex;
     bool exit = false;
-    int duration = 0;
+    int64_t duration = 0;
     bool supportMediacodec = false;
     const AVBitStreamFilter *bsFilter = NULL;
     int tempIndex = 0;
